feat: rename session strategy (#105)
BREAKING CHANGE:

We have always supported both database-persisted and "stateless" sessions, where the session is saved as a JWT in a session cookie. To change between these there was a `jwt` boolean flag the user could toggle, but I think the name appeared too much in the config and might have created confusion about how to use it or what it did. This PR introduces the following change:

```diff
session: {
-  jwt: true, // or false, which is the default value if not set when using an adapter
+ strategy: "database" // or "jwt", which is the default value if no adapter set
}
```

Important note, you can use `strategy: "jwt"`, even if you use an Adapter, but not the other way (makes no sense to use `strategy: "database"` when you don't have a database set up.)