docs(refresh-token): update jwt handler to read from correct field
The new field, `expires_at` is actually in seconds and represents a time in the future, not the amount of time before a token expires. 

The docs, as written before this change, were extended the lifespan of the original token well beyond what was valid causing the refresh to never be triggered.
This change fixes that by removing the addition of `Date.now()` and simply converting the time we get into milliseconds.